requirements.txt
djangorestframework==3.11.2
Django==3.1.8
PyJWT==1.7.1
djangorestframework-jwt
requests
